Move getCtaLabel to useStoreItemInfo hook for reusability
