+++ client/my-sites/plans/jetpack-plans/product-store/hooks/use-store-item-info.tsx
@@ -4,13 +4,16 @@ import {
 	TERM_ANNUALLY,
 	TERM_MONTHLY,
 } from '@automattic/calypso-products';
+import { useTranslate } from 'i18n-calypso';
 import { useCallback, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import isSupersedingJetpackItem from 'calypso/../packages/calypso-products/src/is-superseding-jetpack-item';
 import { getPurchaseByProductSlug } from 'calypso/lib/purchases/utils';
+import OwnerInfo from 'calypso/me/purchases/purchase-item/owner-info';
 import { getSitePurchases } from 'calypso/state/purchases/selectors';
 import { useIsUserPurchaseOwner } from 'calypso/state/purchases/utils';
 import { getSitePlan, getSiteProducts } from 'calypso/state/sites/selectors';
+import productButtonLabel from '../../product-card/product-button-label';
 import { SelectorProduct } from '../../types';
 import { UseStoreItemInfoProps } from '../types';
 
@@ -26,6 +29,7 @@ export const useStoreItemInfo = ( {
 	const siteProducts = useSelector( ( state ) => getSiteProducts( state, siteId ) );
 	const purchases = useSelector( ( state ) => getSitePurchases( state, siteId ) );
 	const isCurrentUserPurchaseOwner = useIsUserPurchaseOwner();
+	const translate = useTranslate();
 
 	// Determine whether product is owned.
 	const isOwned = useCallback(
@@ -123,9 +127,39 @@ export const useStoreItemInfo = ( {
 		[ getPurchase, isCurrentUserPurchaseOwner ]
 	);
 
+	const getCtaLabel = useCallback(
+		( item: SelectorProduct ) => {
+			const ctaLabel = productButtonLabel( {
+				product: item,
+				isOwned: isOwned( item ),
+				isUpgradeableToYearly: isUpgradeableToYearly( item ),
+				isDeprecated: isDeprecated( item ),
+				isSuperseded: isSuperseded( item ),
+				currentPlan: sitePlan,
+				fallbackLabel: translate( 'Get' ),
+			} );
+
+			const purchase = getPurchase( item );
+
+			if ( ! purchase ) {
+				return ctaLabel;
+			}
+
+			return (
+				<>
+					{ ctaLabel }
+					&nbsp;
+					<OwnerInfo purchase={ purchase } />
+				</>
+			);
+		},
+		[ getPurchase, isOwned, isSuperseded, isUpgradeableToYearly, sitePlan, translate ]
+	);
+
 	return useMemo(
 		() => ( {
 			getCheckoutURL,
+			getCtaLabel,
 			getOnClickPurchase,
 			getPurchase,
 			isDeprecated,
@@ -140,6 +174,7 @@ export const useStoreItemInfo = ( {
 		} ),
 		[
 			getCheckoutURL,
+			getCtaLabel,
 			getOnClickPurchase,
 			getPurchase,
 			isIncludedInPlan,
